refactor(embedding): moving nomic embedder under embedder_ollama and refactor embedder as base class
