+++ src/researcher/embedding/__init__.py
@@ -1,7 +1,7 @@
-from .embedder import NomicEmbedder
 from .chunkers import simple_chunk
 from .search import EmbeddingSearcher
 from .index_builder import build_annoy_index
+from .embedder import Embedder
 
-__all__ = ["NomicEmbedder", "simple_chunk", "EmbeddingSearcher"
+__all__ = ["Embedder", "simple_chunk", "EmbeddingSearcher"
            , "build_annoy_index"]
